test_par_gj.py: make pytest-friendly and actually run it
